Change sort to return integer instead of boolean

+++ src/Http/RateLimit/Handler.php
@@ -108,7 +108,7 @@ class Handler
             // limiting will not be imposed for the request.
 
             $this->throttle = $this->getMatchingThrottles()->sort(function ($a, $b) {
-                return $a->getLimit() < $b->getLimit();
+                return (int) ($a->getLimit() < $b->getLimit());
             })->first();
         }
